updated links
Removed the deleted facebook account, and removed mention of my time as a moderator for freeCodeCamp. I'm still proud of those things, but how do you brag without sending people to Facebook, something I cannot do. Kept the link and mention of membership at freeCodeCamp, because fcc is important. They do good work over there. Added the SmartBoard to my portfolio. It's not finished, and it's issues are numerous, but it demonstrates a few things with its ridiculous over-engineered design, highlighting work I've done with canvas in particular.